Make Badge and Tooltip a new type of component called wrappers
