+++ src/panel_material_ui/__init__.py
@@ -7,5 +7,6 @@ from .pane import *  # noqa
 from .template import *  # noqa
 from .theme import MaterialDesign  # noqa
 from .widgets import *  # noqa
+from .wrappers import *  # noqa
 
 import panel_material_ui.param  # noqa
