+++ src/panel_material_ui/pane/base.py
@@ -6,7 +6,6 @@ from typing import Any
 import param
 from panel.links import Callback
 from panel.pane import HoloViews, Markdown
-from panel.viewable import Child
 
 from ..base import COLORS, MaterialComponent, ThemedTransform, TooltipTransform
 from ..widgets import DatetimeInput, DiscreteSlider, EditableFloatSlider, EditableIntSlider, FloatSlider, IntSlider, Select
@@ -124,89 +123,6 @@ class ClickablePaneBase(MaterialPaneBase):
             if val is not None:
                 callbacks[k] = val
         return Callback(self, code=callbacks, args=args)
-
-
-class Badge(MaterialComponent):
-    """
-    The `Badge` generates a small badge to the top-right (by default)
-    of its child element. Badges are commonly used to display
-    notification counts, status indicators, or short labels overlaid
-    on icons, avatars, or buttons.
-
-    :References:
-
-    - https://panel-material-ui.holoviz.org/reference/panes/Badge.html
-    - https://mui.com/material-ui/react-badge/
-
-    :Example:
-
-    >>> Badge(IconButton(icon="mail"), badge_content=4, color="primary")
-    """
-
-    anchor_origin = param.Dict(default=None, doc="""
-        The anchor position of the badge. Accepts a dictionary with
-        keys 'vertical' ('top' or 'bottom') and 'horizontal'
-        ('left' or 'right').""")
-
-    badge_content = param.Parameter(default=0, doc="""
-        The content rendered within the badge. Typically an integer
-        count but can be a short string.""")
-
-    color = param.Selector(default="primary", objects=COLORS, doc="""
-        The color of the badge.""")
-
-    max = param.Integer(default=99, bounds=(0, None), doc="""
-        Maximum count to display. Values above this show as
-        'max+' (e.g. '99+').""")
-
-    object = Child(doc="""
-        The child component to wrap with the badge.""")
-
-    offset = param.XYCoordinate(default=None, doc="""
-        The (x, y) pixel offset of the badge from its anchor point on
-        the object. Positive x shifts the badge right, positive y down.""")
-
-    overlap = param.Selector(default="rectangular", objects=["rectangular", "circular"], doc="""
-        Wrapped shape the badge should overlap.""")
-
-    show_zero = param.Boolean(default=False, doc="""
-        Whether to display the badge when badge_content is zero.""")
-
-    variant = param.Selector(default="standard", objects=["dot", "standard"], doc="""
-        The variant of the badge. Use 'dot' for a small dot
-        indicator without content.""")
-
-    _esm_base = "Badge.jsx"
-
-    def __init__(self, object=None, **params):
-        if object is not None:
-            params["object"] = object
-        # When the user doesn't set the Badge's own margin it inherits the
-        # child's margin, so wrapping in a Badge is transparent to layout (the
-        # child's margin is zeroed in the frontend so the badge still hugs the
-        # child's edge). An explicit Badge margin always wins.
-        margin_explicit = "margin" in params
-        super().__init__(**params)
-        self._margin_explicit = margin_explicit
-        self._child_margin_watcher = None
-        self._sync_child_margin()
-        self.param.watch(self._sync_child_margin, "object")
-
-    def _sync_child_margin(self, *_):
-        if self._child_margin_watcher is not None:
-            # On an object swap self.object already points at the new child, so
-            # unwatch via the stored reference rather than self.object.
-            watcher, previous = self._child_margin_watcher
-            previous.param.unwatch(watcher)
-            self._child_margin_watcher = None
-        if self._margin_explicit:
-            return
-        obj = self.object
-        if not isinstance(obj, param.Parameterized) or "margin" not in obj.param:
-            return
-        self.margin = obj.margin
-        watcher = obj.param.watch(lambda *_: setattr(self, "margin", obj.margin), "margin")
-        self._child_margin_watcher = (watcher, obj)
 
 
 class Avatar(ClickablePaneBase):
@@ -393,71 +309,6 @@ class Skeleton(MaterialPaneBase):
     )
 
     _esm_base = "Skeleton.jsx"
-
-
-class Tooltip(MaterialComponent):
-    """
-    The `Tooltip` displays informative text when users hover over, focus
-    on, or tap a child element. It wraps a single child component and
-    shows a configurable tooltip label.
-
-    :References:
-
-    - https://panel-material-ui.holoviz.org/reference/panes/Tooltip.html
-    - https://mui.com/material-ui/react-tooltip/
-
-    :Example:
-
-    >>> Tooltip(Button(label="Delete"), title="Remove this item")
-    """
-
-    arrow = param.Boolean(default=False, doc="""
-        Whether the tooltip has an arrow indicating the element it
-        refers to.""")
-
-    describe_child = param.Boolean(default=False, doc="""
-        Whether the tooltip acts as an accessible description rather
-        than a label. Use when the child already has a visible label
-        and the tooltip provides supplementary information.""")
-
-    enter_delay = param.Integer(default=100, bounds=(0, None), doc="""
-        The number of milliseconds to wait before showing the tooltip.
-        This can help avoid tooltips appearing on quick mouse passes.""")
-
-    follow_cursor = param.Boolean(default=False, doc="""
-        Whether the tooltip follows the cursor position.""")
-
-    leave_delay = param.Integer(default=0, bounds=(0, None), doc="""
-        The number of milliseconds to wait before hiding the tooltip.""")
-
-    object = Child(doc="""
-        The child component to wrap with the tooltip.""")
-
-    open = param.Boolean(default=None, allow_None=True, doc="""
-        Explicitly control whether the tooltip is open. When None,
-        the tooltip is managed automatically on hover/focus. Set to
-        True or False for programmatic control.""")
-
-    placement = param.Selector(
-        default="bottom",
-        objects=[
-            "bottom-end", "bottom-start", "bottom",
-            "left-end", "left-start", "left",
-            "right-end", "right-start", "right",
-            "top-end", "top-start", "top",
-        ],
-        doc="""
-        The placement of the tooltip relative to the child element.""")
-
-    title = param.String(default="", doc="""
-        The text to display inside the tooltip.""")
-
-    _esm_base = "Tooltip.jsx"
-
-    def __init__(self, object=None, **params):
-        if object is not None:
-            params["object"] = object
-        super().__init__(**params)
 
 
 class Typography(MaterialPaneBase, Markdown):
