+++ src/panel_material_ui/wrappers/__init__.py
@@ -1,0 +1,1 @@
+from .base import *  # noqa
