+++ tests/ui/wrappers/test_tooltip.py
@@ -3,7 +3,7 @@ import pytest
 pytest.importorskip('playwright')
 
 from panel.tests.util import serve_component
-from panel_material_ui.pane import Tooltip
+from panel_material_ui.wrappers import Tooltip
 from panel_material_ui.widgets import Button
 from playwright.sync_api import expect
